requirements.txt
PyPDF2 
openai
pinecone
fastapi
python-dotenv